2008-09-05 18:12 UTC+0200 Viktor Szakats (harbour.01 syenar hu)    * contrib/hbapollo/apollo.c    * contrib/hbclipsm/gauge.c    * contrib/hbct/atadjust.c    * contrib/hbct/atnum.c    * contrib/hbct/atrepl.c    * contrib/hbct/ctstr.c    * contrib/hbct/ctwfunc.c    * contrib/hbct/dattime2.c    * contrib/hbct/dattime3.c    * contrib/hbct/files.c    * contrib/hbct/pack.c    * contrib/hbct/print.c    * contrib/hbct/setrc.c    * contrib/hbfbird/firebird.c    * contrib/hbhpdf/harupdf.c    * contrib/hbmisc/hb_f.c    * contrib/hbmzip/hbmzip.c    * contrib/hbnf/descendn.c    * contrib/hbnf/dispc.c    * contrib/hbnf/ftattr.c    * contrib/hbnf/ftshadow.c    * contrib/hbnf/fttext.c    * contrib/hbodbc/odbc.c    * contrib/hbole/ole2.c    * contrib/hbw32/dllcall.c    * contrib/hbw32/tprinter.c    * contrib/rddads/ads1.c    * contrib/xhb/cstructc.c    * contrib/xhb/hbxml.c    * contrib/xhb/xhbmsgs.c    * utils/hbmake/hbmgauge.c      ! Fixed some MSVS 2008 -W4 warnings.      ; There are quite some more.
   * contrib/hbhpdf/harupdf.c
     ! HPDF_ADDPAGELABEL() accepting first parameter wrongly.

   * contrib/hbfbird/firebird.c
     ! FBQUERY() potential GPF when using value from freed memory.

   * doc/whatsnew.txt
     * Updated for latest fixes.

+++ harbour/contrib/hbct/dattime3.c
@@ -294,10 +294,10 @@ HB_FUNC( SETTIME )
 #if defined(HB_OS_WIN_32)
       SYSTEMTIME st;
       GetLocalTime( &st );
-      st.wHour         = iTime[0];
-      st.wMinute       = iTime[1];
-      st.wSecond       = iTime[2];
-      st.wMilliseconds = iTime[3] * 10;
+      st.wHour         = ( WORD ) iTime[0];
+      st.wMinute       = ( WORD ) iTime[1];
+      st.wSecond       = ( WORD ) iTime[2];
+      st.wMilliseconds = ( WORD ) iTime[3] * 10;
       fResult = SetLocalTime( &st );
 #elif defined( HB_OS_LINUX ) && !defined( __WATCOMC__ )
 /* stime exists only in SVr4, SVID, X/OPEN and Linux */
@@ -380,10 +380,10 @@ HB_FUNC( SETDATE )
 #if defined(HB_OS_WIN_32)
          SYSTEMTIME st;
          GetLocalTime( &st );
-         st.wYear      = iYear;
-         st.wMonth     = iMonth;
-         st.wDay       = iDay;
-         st.wDayOfWeek = hb_dateJulianDOW( lDate );
+         st.wYear      = ( WORD ) iYear;
+         st.wMonth     = ( WORD ) iMonth;
+         st.wDay       = ( WORD ) iDay;
+         st.wDayOfWeek = ( WORD ) hb_dateJulianDOW( lDate );
          fResult = SetLocalTime( &st );
 #elif defined( HB_OS_LINUX ) && !defined( __WATCOMC__ )
 /* stime exists only in SVr4, SVID, X/OPEN and Linux */
